requirements.txt

pandas==2.0.3
pytrends==4.9.2
google-cloud-bigquery==3.11.4
google-auth==2.23.0